refactor(settings): move power settings to Power
Group auto-suspend, auto power off, and sleep cover under a dedicated
Power category ahead of soft-suspend settings.
