+++ docs/po/messages.pot
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: Cadmus Documentation\n"
-"POT-Creation-Date: 2026-08-08T10:11:18Z\n"
+"POT-Creation-Date: 2026-08-04T11:45:36+02:00\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
@@ -82,111 +82,103 @@ msgstr ""
 msgid "Importing"
 msgstr ""
 
-#: src/SUMMARY.md:22 src/wifi.md:1
-msgid "WiFi"
-msgstr ""
-
-#: src/SUMMARY.md:23 src/time-sync.md:1
+#: src/SUMMARY.md:22 src/time-sync.md:1
 msgid "Automatic Time Syncing"
 msgstr ""
 
-#: src/SUMMARY.md:24 src/auto-frontlight.md:1
+#: src/SUMMARY.md:23 src/auto-frontlight.md:1
 msgid "Automatic Frontlight"
 msgstr ""
 
-#: src/SUMMARY.md:25 src/database-backup.md:1
+#: src/SUMMARY.md:24 src/database-backup.md:1
 msgid "Database Backup"
 msgstr ""
 
-#: src/SUMMARY.md:26 src/installation/ota.md:86 src/troubleshooting/index.md:1
+#: src/SUMMARY.md:25 src/installation/ota.md:86 src/troubleshooting/index.md:1
 msgid "Troubleshooting"
 msgstr ""
 
-#: src/SUMMARY.md:27
+#: src/SUMMARY.md:26
 msgid "CHANGELOG"
 msgstr ""
 
-#: src/SUMMARY.md:31
+#: src/SUMMARY.md:30
 msgid "Contributor Guide"
 msgstr ""
 
-#: src/SUMMARY.md:33
+#: src/SUMMARY.md:32
 msgid "Development Environment"
 msgstr ""
 
-#: src/SUMMARY.md:34
+#: src/SUMMARY.md:33
 msgid "SSH"
 msgstr ""
 
-#: src/SUMMARY.md:35
+#: src/SUMMARY.md:34
 msgid "Code Style and Linting"
 msgstr ""
 
-#: src/SUMMARY.md:36
+#: src/SUMMARY.md:35
 msgid "Event System"
 msgstr ""
 
-#: src/SUMMARY.md:37
-msgid "WiFi Leases"
-msgstr ""
-
-#: src/SUMMARY.md:38 src/settings/index.md:839
+#: src/SUMMARY.md:36 src/settings/index.md:839
 msgid "Telemetry"
 msgstr ""
 
-#: src/SUMMARY.md:39
+#: src/SUMMARY.md:37
 msgid "Logging"
 msgstr ""
 
-#: src/SUMMARY.md:40
+#: src/SUMMARY.md:38
 msgid "Tracing"
 msgstr ""
 
-#: src/SUMMARY.md:41
+#: src/SUMMARY.md:39
 msgid "Profiling"
 msgstr ""
 
-#: src/SUMMARY.md:42
+#: src/SUMMARY.md:40
 msgid "Website"
 msgstr ""
 
-#: src/SUMMARY.md:43
+#: src/SUMMARY.md:41
 msgid "User guide"
 msgstr ""
 
-#: src/SUMMARY.md:44
+#: src/SUMMARY.md:42
 msgid "Translations"
 msgstr ""
 
-#: src/SUMMARY.md:45
+#: src/SUMMARY.md:43
 msgid "For developers"
 msgstr ""
 
-#: src/SUMMARY.md:46
+#: src/SUMMARY.md:44
 msgid "SQLite & SQLx"
 msgstr ""
 
-#: src/SUMMARY.md:47
+#: src/SUMMARY.md:45
 msgid "Library Database"
 msgstr ""
 
-#: src/SUMMARY.md:48
+#: src/SUMMARY.md:46
 msgid "Runtime Migrations"
 msgstr ""
 
-#: src/SUMMARY.md:52
+#: src/SUMMARY.md:50
 msgid "Investigations"
 msgstr ""
 
-#: src/SUMMARY.md:54
+#: src/SUMMARY.md:52
 msgid "Intro"
 msgstr ""
 
-#: src/SUMMARY.md:55
+#: src/SUMMARY.md:53
 msgid "DHCP IP Address Changes on WiFi Toggle"
 msgstr ""
 
-#: src/SUMMARY.md:56
+#: src/SUMMARY.md:54
 msgid "Reading IP and ESSID via dhcpcd-dbus"
 msgstr ""
 
@@ -2474,9 +2466,11 @@ msgid "Categories"
 msgstr ""
 
 #: src/ui/settings/index.md:12
-msgid ""
-"**General** — language, sleep, auto-suspend, WiFi idle timeout, button "
-"layout, startup mode"
+msgid "**General** — language, button layout, Wi‑Fi idle timeout, startup mode"
+msgstr ""
+
+#: src/ui/settings/index.md:13
+msgid "**Power** — auto-suspend, auto power off, sleep cover"
 msgstr ""
 
 #: src/ui/settings/index.md:14
@@ -2827,47 +2821,6 @@ msgid ""
 "This bypasses the incremental import cache and re-fingerprints all files in "
 "your library directories. Be aware that this can take time and drain the "
 "battery for large libraries, so keep your device plugged in while it runs."
-msgstr ""
-
-#: src/wifi.md:3
-msgid ""
-"Cadmus can turn WiFi on when something needs the network, then turn it off "
-"again when it is idle."
-msgstr ""
-
-#: src/wifi.md:6
-msgid "Modes"
-msgstr ""
-
-#: src/wifi.md:8
-msgid "Open the **WiFi** control in the top menu and choose one of:"
-msgstr ""
-
-#: src/wifi.md:10
-msgid ""
-"**Off** — WiFi stays off. Features that need the network cannot turn it on "
-"automatically; you must turn WiFi on yourself beforehand."
-msgstr ""
-
-#: src/wifi.md:12
-msgid "**Always On** — WiFi stays on."
-msgstr ""
-
-#: src/wifi.md:13
-msgid ""
-"**Auto** — Cadmus turns WiFi on when something needs the network, then "
-"powers it down after a period of inactivity."
-msgstr ""
-
-#: src/wifi.md:16
-msgid "Idle timeout"
-msgstr ""
-
-#: src/wifi.md:18
-msgid ""
-"When WiFi is set to **Auto**, Cadmus waits a short time after the last "
-"network use before powering the radio down. Change that delay in the "
-"[settings editor](ui/settings/index.md)."
 msgstr ""
 
 #: src/time-sync.md:3
