+++ docs/src/ui/settings/index.md
@@ -9,8 +9,8 @@ Settings are organised into tabs — tap a category to open it.
 
 ## Categories
 
-- **General** — language, sleep, auto-suspend, WiFi idle timeout, button
-  layout, startup mode
+- **General** — language, button layout, Wi‑Fi idle timeout, startup mode
+- **Power** — auto-suspend, auto power off, sleep cover
 - **Reader** — what happens when you finish a book
 - **Libraries** — add, edit, or remove your book libraries
 - **Dictionaries** — download and manage offline dictionaries
